requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psutil==5.4.5
python-gnupg==0.4.3
python3-memcached==1.51
redis==2.10.6
requests==2.18.4
six==1.11.0
urllib3==1.22
uWSGI==2.0.17.1
Werkzeug==0.14.1
